split out py arena
